refactor: 拆分 status_monitor 为 jobs/commands
